requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
dnspython=2.4.2
python=3.11
tqdm
